requirements.txt
psycopg2-binary==2.8.3
Django==3.0.3
djangorestframework==3.10.3
django-cors-headers==3.2.1
pylint-django==2.0.14
# astroid==2.0.4
# pylint==2.1.1
coverage==5.1
